requirements.txt
celery==5.4.0
Django==5.1
django-celery-results==2.5.1
redis==5.0.8
requests==2.32.3
pillow==10.4.0
whitenoise==6.6.0
